Rename 'CardSelector' -> 'Selector'
